// examples/nimbus/agent-worker/nimbus/_generated/dataModel.d.ts
// Generated by @nimbus/codegen. Do not edit by hand.
import type { GenericId } from "@nimbus/nimbus/values";

export type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export type TableNames = "jobs";

export type Id<TableName extends string> = GenericId<TableName>;

type DocumentByTable = {
  "jobs": {
    _id: Id<"jobs">;
    _creationTime: number;
    _updateTime: number;
    "label": string;
    "status": string;
    "createdAt": number;
    "completedAt": number | undefined;
  };
};

type IndexNamesByTable = {
  "jobs": "by_status";
};

export type Doc<TableName extends TableNames> = DocumentByTable[TableName];

export type DataModel = {
  [TableName in TableNames]: {
    document: DocumentByTable[TableName];
    fieldPaths: keyof DocumentByTable[TableName];
    indexes: IndexNamesByTable[TableName];
  };
};